Remove unused tasks and update instructions
