--- test/phantom_hooks.js
@@ -1,17 +1,0 @@
-/* eslint-disable no-console */
-
-module.exports = {
-  afterEnd: function(runner) {
-    var fs = require('fs');
-    var coverage = runner.page.evaluate(function() {
-      return window.__coverage__;
-    });
-
-    if (coverage) {
-      console.log('Writing coverage to coverage/coverage.json');
-      fs.write('coverage/coverage.json', JSON.stringify(coverage), 'w');
-    } else {
-      console.log('No coverage data generated');
-    }
-  }
-};
